Add additional stats to the contribution page
New query is executed at the page to display some stats from our
aggregation table. This should not have performance impact as the table
is quite small and ideally the database server will cache the subquery
results so that it can be immediately used by the table_sql to display
actual data.

The patch also simplifies table rows counting where we do not need to
join the user table.